continuing to organize packages/files. trying to clean up the methods that cURL for new issues
